# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: submissions
    description: Volunteer hour submission operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getAuthStatus
      tags: [auth]
      summary: Get current auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, all]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Create a new submission (multipart form)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /submissions/stats:
    get:
      operationId: getSubmissionStats
      tags: [submissions]
      summary: Get submission statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionStats"

  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a single submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /submissions/upload-token:
    get:
      operationId: getUploadToken
      tags: [submissions]
      summary: Get an upload URL/token for file upload
      responses:
        "200":
          description: Upload info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        username:
          type: string
      required:
        - success
        - username

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: ["string", "null"]
      required:
        - authenticated
        - username

    FileUploadResponse:
      type: object
      properties:
        fileUrl:
          type: string
        fileName:
          type: string
      required:
        - fileUrl
        - fileName

    Submission:
      type: object
      properties:
        id:
          type: integer
        submissionId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        studentId:
          type: string
        graduationYear:
          type: integer
        email:
          type: ["string", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        extractedOrg:
          type: ["string", "null"]
        extractedHours:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - submissionId
        - firstName
        - lastName
        - studentId
        - graduationYear
        - email
        - fileUrl
        - fileName
        - status
        - notes
        - extractedOrg
        - extractedHours
        - createdAt
        - updatedAt

    SubmissionInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        studentId:
          type: string
        graduationYear:
          type: integer
        email:
          type: ["string", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
      required:
        - firstName
        - lastName
        - studentId
        - graduationYear
        - fileUrl
        - fileName

    SubmissionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]

    SubmissionStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
      required:
        - total
        - pending
        - approved
        - rejected