# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionStatus } from "./submissionStatus";

export interface Submission {
  id: number;
  submissionId: string;
  firstName: string;
  lastName: string;
  studentId: string;
  graduationYear: number;
  /** @nullable */
  email: string | null;
  fileUrl: string;
  fileName: string;
  status: SubmissionStatus;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  extractedOrg: string | null;
  /** @nullable */
  extractedHours: string | null;
  createdAt: Date;
  updatedAt: Date;
}