# lib/api-zod/src/generated/types/submissionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionInput {
  firstName: string;
  lastName: string;
  studentId: string;
  graduationYear: number;
  /** @nullable */
  email?: string | null;
  fileUrl: string;
  fileName: string;
}