# lib/api-zod/src/generated/types/submissionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionUpdateStatus } from "./submissionUpdateStatus";

export interface SubmissionUpdate {
  status?: SubmissionUpdateStatus;
  /** @nullable */
  notes?: string | null;
}